AssignedTo field added to Result Model
AssignedTo will contain a string with the name of the QA in charge of
running the test case for a result.
From ExecuteTestRun/Index, the user can add this assignee trough a modal
dialog (#modalAssignTo) displayed on a partial view "SetAssignee".
Seeds with Results were updated with data for this field.