Project CreatedOn and LastModified fix
There is no need to set CreatedOn manually, DB context does so.
SetValues method for CurrentValues is used to update an entity:

SetValues copies the properties of the <entityToUpdate> to the
properties of the entity which has the same key and is already attached
to the context if the properties have the same names. If the property
values are different it marks the property as Modified.