Create result is now triggered by an Action Link
Dropdowns were dismissed cause we had the case where a new result with
the same actual state of the dropdown needed to be created. This meant a
non-change on the selected id, therefore no function triggered.

The routing is not needed anymore.

Changed rows divs per a Table.

Deleted an old ViewModel too.